refactor main.js
